+++ include/endstone/metrics/simple_bar_chart.h
@@ -34,17 +34,17 @@ public:
     {
     }
 
-    [[nodiscard]] std::optional<ChartData> getChartData() override
+    [[nodiscard]] nlohmann::json getChartData() override
     {
         const auto map_values = get_values_();
         if (!map_values || map_values->empty()) {
-            return std::nullopt;
+            return nullptr;
         }
-        BarValues values;
+        auto values = nlohmann::json::object();
         for (const auto &[key, value] : *map_values) {
-            values.emplace(key, std::vector{value});
+            values[key] = nlohmann::json::array({value});
         }
-        return ChartData{{"values", std::move(values)}};
+        return nlohmann::json{{"values", std::move(values)}};
     }
 
 private: